refactor: split page into multiple components
