+++ generator/nextjs/template/src/pages/iota-callback/iota-callback-error.tsx
@@ -1,0 +1,22 @@
+export default function IotaCallbackPageError({
+  errorMessage,
+  featureAvailable,
+}: {
+  featureAvailable: boolean;
+  errorMessage: string | null;
+}) {
+  if (!featureAvailable) {
+    return (
+      <div>
+        Feature not available. Please set your Personal Access Token in your
+        environment secrets.
+      </div>
+    );
+  }
+
+  return (
+    <div>
+      <h1>{errorMessage}</h1>
+    </div>
+  );
+}
